add proto files and optimize code standard

+++ ChatServer/NodeNet/ChatNodeServer.cpp
@@ -37,7 +37,7 @@ bool ChatNodeServer::SendPackToGame(const int msg_id, google::protobuf::Message*
 	return true;
 }
 
-void ChatNodeServer::OnGameRouteBack(socket_t sock_fd, const int msg_id, const char* msg, const uint32_t msg_len)
+void ChatNodeServer::OnGameRouteBack(socket_t sock_fd, const int msg_id, const char* msg, const size_t msg_len)
 {
 }
