+++ ChatServer/NodeNet/ChatNodeServer.h
@@ -14,5 +14,5 @@ public:
 	bool SendPackToGame(const int msg_id, google::protobuf::Message* msg);
 
 	//back msg to client from other server
-	void OnGameRouteBack(socket_t sock_fd, const int msg_id, const char* msg, const uint32_t msg_len);
+	void OnGameRouteBack(socket_t sock_fd, const int msg_id, const char* msg, const size_t msg_len);
 };
