+++ DBServer/NodeNet/DBNodeServer.cpp
@@ -18,7 +18,7 @@ bool DBNodeServer::InitHelper()
 	return true;
 }
 
-void DBNodeServer::OtherMessage(const socket_t sock_fd, const int msg_id, const char* msg, const uint32_t msg_len)
+void DBNodeServer::OtherMessage(const socket_t sock_fd, const int msg_id, const char* msg, const size_t msg_len)
 {
 
 }
