+++ GateServer/NodeNet/GatePlayerServer.h
@@ -15,7 +15,7 @@ public:
 	void OnClientConnected(const socket_t sock_fd);
 	void OnClientDisconnect(const socket_t sock_fd);
 
-	void OnOtherMessage(const socket_t sock_fd, const int msg_id, const char* msg, const uint32_t msg_len);
+	void OnOtherMessage(const socket_t sock_fd, const int msg_id, const char* msg, const size_t msg_len);
 	void SentToClient(const int msg_id, const std::string& msg, const socket_t sock_fd);
 	void SentToClient(const int msg_id, google::protobuf::Message* msg, const socket_t sock_fd);
 	void SentToAllClient(const int msg_id, const std::string& msg);
